more changes for refactor
