+++ fringe/src/fringe/templates/memory/MemPrimitives.scala
@@ -13,70 +13,6 @@ import fringe.utils.HVec
 import fringe.utils.{getRetimed, log2Up}
 import fringe.utils.implicits._
 
-class R_XBar(val port_width: Int, val ofs_width:Int, val bank_width:List[Int]) extends Bundle {
-  val banks = HVec.tabulate(port_width*bank_width.length){i => UInt(bank_width(i%bank_width.length).W)}
-  val ofs = Vec(port_width, UInt(ofs_width.W))
-  val en = Vec(port_width, Bool())
-  val backpressure = Vec(port_width, Bool())
-
-  def connectLane(lhs_lane: Int, rhs_lane: Int, rhs_port: R_XBar, f: Bool): Unit = {
-    ofs(lhs_lane) := rhs_port.ofs(rhs_lane)
-    bank_width.length.indices[Unit]{i => banks(i + lhs_lane*bank_width.length) := rhs_port.banks(i + rhs_lane*bank_width.length)}
-    en(lhs_lane) := rhs_port.en(rhs_lane)
-    backpressure(lhs_lane) := f
-  }
-
-  def forwardLane(lane: Int, rhs: R_XBar): Unit = {
-    ofs(lane) := rhs.ofs(lane)
-    bank_width.length.indices[Unit]{i => banks(i + lane*bank_width.length) := rhs.banks(i + lane*bank_width.length)}
-    en(lane) := rhs.en(lane)
-    backpressure(lane) := rhs.backpressure(lane)
-  }
-
-  override def cloneType = (new R_XBar(port_width, ofs_width, bank_width)).asInstanceOf[this.type] // See chisel3 bug 358
-}
-
-class W_XBar(val port_width: Int, val ofs_width:Int, val bank_width:List[Int], val data_width:Int) extends Bundle {
-  val banks = HVec.tabulate(port_width*bank_width.length){i => UInt(bank_width(i%bank_width.length).W)}
-  val ofs = Vec(port_width, UInt(ofs_width.W))
-  val data = Vec(port_width, UInt(data_width.W))
-  val reset = Vec(port_width, Bool()) // For FF
-  val init = Vec(port_width, UInt(data_width.W)) // For FF
-  val shiftEn = Vec(port_width, Bool()) // For ShiftRegFile
-  val en = Vec(port_width, Bool())
-
-  override def cloneType = (new W_XBar(port_width, ofs_width, bank_width, data_width)).asInstanceOf[this.type] // See chisel3 bug 358
-}
-
-class R_Direct(val port_width:Int, val ofs_width:Int, val banks:List[List[Int]]) extends Bundle {
-  val ofs = Vec(port_width, UInt(ofs_width.W))
-  val en = Vec(port_width, Bool())
-  val backpressure = Vec(port_width, Bool())
-
-  def connectLane(lhs_lane: Int, rhs_lane: Int, rhs_port: R_Direct, f: Bool): Unit = {
-    ofs(lhs_lane) := rhs_port.ofs(rhs_lane)
-    en(lhs_lane) := rhs_port.en(rhs_lane)
-    backpressure(lhs_lane) := f
-  }
-
-  def forwardLane(lane: Int, rhs: R_Direct): Unit = {
-    ofs(lane) := rhs.ofs(lane)
-    en(lane) := rhs.en(lane)
-    backpressure(lane) := rhs.backpressure(lane)
-  }
-
-  override def cloneType = (new R_Direct(port_width, ofs_width, banks)).asInstanceOf[this.type] // See chisel3 bug 358
-}
-
-class W_Direct(val port_width: Int, val ofs_width:Int, val banks:List[List[Int]], val data_width:Int) extends Bundle {
-  val ofs = Vec(port_width, UInt(ofs_width.W))
-  val data = Vec(port_width, UInt(data_width.W))
-  val shiftEn = Vec(port_width, Bool()) // For ShiftRegFile
-  val en = Vec(port_width, Bool())
-
-  override def cloneType = (new W_Direct(port_width, ofs_width, banks, data_width)).asInstanceOf[this.type] // See chisel3 bug 358
-}
-
 
 abstract class MemPrimitive(val p: MemParams) extends Module {
   val io = p.iface match {
