+++ fringe/src/fringe/templates/memory/NBuffers.scala
@@ -68,234 +68,6 @@ class NBufController(numBufs: Int, portsWithWriter: List[Int]) extends Module {
 
 }
 
-case class NBufParams(
-  val mem: MemType,
-  val logicalDims: List[Int],
-  val numBufs: Int,
-  val bitWidth: Int,
-  val banks: List[Int],
-  val strides: List[Int],
-  val xBarWMux: NBufXMap,
-  val xBarRMux: NBufXMap, // bufferPort -> (muxPort -> accessPar)
-  val directWMux: NBufDMap,
-  val directRMux: NBufDMap,  // bufferPort -> (muxPort -> List(banks, banks, ...))
-  val broadcastWMux: XMap,
-  val broadcastRMux: XMap,  // Assume broadcasts are XBar
-  val bankingMode: BankingMode,
-  val inits: Option[List[Double]] = None,
-  val syncMem: Boolean = false,
-  val fracBits: Int = 0,
-  val numActives: Int = 1,
-  val myName: String = "NBuf"
-) {
-  def hasXBarW = xBarWMux.accessPars.sum > 0
-  def hasXBarR = xBarRMux.accessPars.sum > 0
-  def numXBarW = xBarWMux.accessPars.sum
-  def numXBarR = xBarRMux.accessPars.sum
-  def numXBarWPorts = xBarWMux.accessPars.size
-  def numXBarRPorts = xBarRMux.accessPars.size
-  def hasDirectW = directWMux.mergeDMaps.accessPars.sum > 0
-  def hasDirectR = directRMux.mergeDMaps.accessPars.sum > 0
-  def numDirectW = directWMux.mergeDMaps.accessPars.sum
-  def numDirectR = directRMux.mergeDMaps.accessPars.sum
-  def numDirectWPorts = directWMux.accessPars.size
-  def numDirectRPorts = directRMux.accessPars.size
-  def hasBroadcastW = broadcastWMux.accessPars.toList.sum > 0
-  def numBroadcastW = broadcastWMux.accessPars.toList.sum
-  def numBroadcastWPorts = broadcastWMux.accessPars.toList.length
-  def hasBroadcastR = broadcastRMux.accessPars.toList.sum > 0
-  def numBroadcastR = broadcastRMux.accessPars.toList.sum
-  def numBroadcastRPorts = broadcastRMux.accessPars.toList.length
-  def totalOutputs = numXBarR + numDirectR + numBroadcastR
-  def defaultDirect = List(List.fill(banks.length)(99))
-  def portsWithWriter = (directWMux.keys ++ xBarWMux.keys).toList.sorted
-  def depth = logicalDims.product + {if (mem == LineBufferType) (numBufs-1)*strides(0)*logicalDims(1) else 0} // Size of memory
-  def N = logicalDims.length // Number of dimensions
-  def ofsWidth = log2Up(depth/banks.product)
-  def banksWidths = banks.map(log2Up(_))
-  def flatXBarWMux = xBarWMux.mergeXMaps
-  def flatXBarRMux = xBarRMux.mergeXMaps
-  def flatDirectWMux = directWMux.mergeDMaps 
-  def flatDirectRMux = directRMux.mergeDMaps 
-  def combinedXBarWMux = flatXBarWMux.merge(broadcastWMux)
-  def combinedXBarRMux = flatXBarRMux.merge(broadcastRMux)
-
-}
-
-class NBufInterface(val p: NBufParams) extends Bundle {
-  val sEn = Vec(p.numBufs, Input(Bool()))
-  val sDone = Vec(p.numBufs, Input(Bool()))
-  val xBarW = HVec(Array.tabulate(1 max p.numXBarWPorts){i => Input(new W_XBar(p.xBarWMux.accessPars.getOr1(i), p.ofsWidth, p.banksWidths, p.bitWidth))})
-  val xBarR = HVec(Array.tabulate(1 max p.numXBarRPorts){i => Input(new R_XBar(p.xBarRMux.accessPars.getOr1(i), p.ofsWidth, p.banksWidths))}) 
-  val directW = HVec(Array.tabulate(1 max p.numDirectWPorts){i => 
-      val dBanks = if (p.hasDirectW) {
-        p.directWMux.toSeq.sortBy(_._1).toMap.values.flatMap(_.toSeq.sortBy(_._1).toMap.values.map(_._1)).toList(i)
-      } else p.defaultDirect
-      Input(new W_Direct(p.directWMux.accessPars.getOr1(i), p.ofsWidth, dBanks, p.bitWidth))
-    })
-  val directR = HVec(Array.tabulate(1 max p.numDirectRPorts){i => 
-      val dBanks = if (p.hasDirectR) {
-        p.directRMux.toSeq.sortBy(_._1).toMap.values.flatMap(_.toSeq.sortBy(_._1).toMap.values.map(_._1)).toList(i)
-      } else p.defaultDirect
-      Input(new R_Direct(p.directRMux.accessPars.getOr1(i), p.ofsWidth, dBanks))
-    })
-  val broadcastW = HVec(Array.tabulate(1 max p.numBroadcastWPorts){i => Input(new W_XBar(p.broadcastWMux.accessPars.getOr1(i), p.ofsWidth, p.banksWidths, p.bitWidth))})
-  val broadcastR = HVec(Array.tabulate(1 max p.numBroadcastRPorts){i => Input(new R_XBar(p.broadcastRMux.accessPars.getOr1(i), p.ofsWidth, p.banksWidths))})
-  val reset = Input(Bool())
-
-  // FIFO Specific
-  val accessActivesIn = Vec(p.numActives, Input(Bool()))
-  val accessActivesOut = Vec(p.numActives, Output(Bool()))
-  val full = Output(Bool())
-  val almostFull = Output(Bool())
-  val empty = Output(Bool())
-  val almostEmpty = Output(Bool())
-  val numel = Output(UInt(32.W))    
-  
-  val output = Vec(1 max p.totalOutputs, Output(UInt(p.bitWidth.W)))  
-
-  def connectLedger(op: NBufInterface)(implicit stack: List[KernelHash]): Unit = {
-    accessActivesOut.zip(op.accessActivesOut).foreach{case (l,r) => r := l}
-    op.full := full
-    op.almostFull := almostFull
-    op.empty := empty
-    op.almostEmpty := almostEmpty
-    op.numel := numel
-    if (Ledger.connections.contains(op.hashCode) && Ledger.connections(op.hashCode).contains(stack.head.hashCode)) {
-      val cxn = Ledger.connections(op.hashCode)(stack.head.hashCode)
-      cxn.xBarR.foreach{case RAddr(p,lane) => xBarR(p).forwardLane(lane, op.xBarR(p))}
-      cxn.xBarW.foreach{p => xBarW(p) <> op.xBarW(p)}
-      cxn.directR.foreach{case RAddr(p,lane) => directR(p).forwardLane(lane, op.directR(p))}
-      cxn.directW.foreach{p => directW(p) <> op.directW(p)}
-      cxn.output.foreach{p => output(p) <> op.output(p)}
-      cxn.broadcastR.foreach{case RAddr(p,lane) => broadcastR(p).forwardLane(lane, op.broadcastR(p))}
-      cxn.broadcastW.foreach{p => broadcastW(p) <> op.broadcastW(p)}
-      cxn.reset.foreach{p => reset <> op.reset}
-    }
-    else this <> op
-  }
-
-  def connectReset(r: Bool)(implicit stack: List[KernelHash]): Unit = {
-    reset := r
-    Ledger.connectReset(this.hashCode, 0)
-  }
-  def connectAccessActivesIn(p: Int, e: Bool)(implicit stack: List[KernelHash]): Unit = {
-    Ledger.connectAccessActivesIn(this.hashCode, p)
-    accessActivesIn(p) := e
-  }
-
-  var usedMuxPorts = List[(String,(Int,Int,Int,Int,Int))]() // Check if the bufferPort, muxPort, muxAddr, lane, castgrp is taken for this connection style (xBar or direct)
-  def connectXBarWPort(wBundle: W_XBar, bufferPort: Int, muxAddr: (Int, Int))(implicit stack: List[KernelHash]): Unit = {
-    assert(p.hasXBarW)
-    assert(!usedMuxPorts.contains(("XBarW", (bufferPort,muxAddr._1,muxAddr._2,0,0))), s"Attempted to connect to XBarW port ($bufferPort,$muxAddr) twice!")
-    usedMuxPorts ::= ("XBarW", (bufferPort,muxAddr._1,muxAddr._2,0,0))
-    val bufferBase = p.xBarWMux.accessParsBelowBufferPort(bufferPort).length
-    val muxBase = p.xBarWMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, muxAddr._2,0).length
-    Ledger.connectXBarW(this.hashCode, bufferBase + muxBase)
-    xBarW(bufferBase + muxBase) := wBundle
-  }
-
-  def connectXBarRPort(rBundle: R_XBar, bufferPort: Int, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean)(implicit stack: List[KernelHash]): Seq[UInt] = {connectXBarRPort(rBundle, bufferPort, muxAddr, castgrps, broadcastids, ignoreCastInfo, true.B)}
-  def connectXBarRPort(rBundle: R_XBar, bufferPort: Int, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean, backpressure: Bool)(implicit stack: List[KernelHash]): Seq[UInt] = {
-    assert(p.hasXBarR)
-    castgrps.zip(broadcastids).zipWithIndex.map{case ((cg, bid), i) => 
-      val castgrp = if (ignoreCastInfo) 0 else cg
-      val effectiveOfs = if (ignoreCastInfo) muxAddr._2 else muxAddr._2 + i
-      val bufferBase = p.xBarRMux.accessParsBelowBufferPort(bufferPort).length
-      val muxBase = p.xBarRMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).length
-      val outputBufferBase = p.xBarRMux.accessParsBelowBufferPort(bufferPort).sum
-      val outputMuxBase = p.xBarRMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).sum
-      val vecId = if (ignoreCastInfo) i else castgrps.take(i).count(_ == castgrp)
-      if (bid == 0) {
-        if (ignoreCastInfo && i == 0) {
-          assert(!usedMuxPorts.contains(("XBarR", (bufferPort,muxAddr._1,effectiveOfs,i,0))), s"Attempted to connect to XBarR port ($bufferPort,$muxAddr) twice!")
-          usedMuxPorts ::= ("XBarR", (bufferPort,muxAddr._1,effectiveOfs,i,0))
-        } else if (!ignoreCastInfo) {
-          assert(!usedMuxPorts.contains(("XBarR", (bufferPort,muxAddr._1,effectiveOfs,i,castgrp))), s"Attempted to connect to XBarR port ($bufferPort,$muxAddr) twice!")
-          usedMuxPorts ::= ("XBarR", (bufferPort,muxAddr._1,effectiveOfs,i,castgrp))
-        }
-        Ledger.connectXBarR(this.hashCode, bufferBase + muxBase, vecId)
-        xBarR(bufferBase + muxBase).connectLane(vecId,i,rBundle, backpressure)
-      }
-      Ledger.connectOutput(this.hashCode, outputBufferBase + outputMuxBase + vecId)
-      output(outputBufferBase + outputMuxBase + vecId)
-    }
-  }
-
-  def connectBroadcastWPort(wBundle: W_XBar, muxAddr: (Int, Int))(implicit stack: List[KernelHash]): Unit = {
-    val muxBase = p.broadcastWMux.accessParsBelowMuxPort(muxAddr._1, muxAddr._2,0).length
-    broadcastW(muxBase) := wBundle
-    Ledger.connectBroadcastW(this.hashCode, muxBase)
-  }
-
-  def connectBroadcastRPort(rBundle: R_XBar, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean)(implicit stack: List[KernelHash]): Seq[UInt] = {connectBroadcastRPort(rBundle, muxAddr, castgrps, broadcastids, ignoreCastInfo, true.B)}
-  def connectBroadcastRPort(rBundle: R_XBar, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean, backpressure: Bool)(implicit stack: List[KernelHash]): Seq[UInt] = {
-    castgrps.zip(broadcastids).zipWithIndex.map{case ((cg, bid), i) => 
-      val castgrp = if (ignoreCastInfo) 0 else cg
-      val effectiveOfs = if (ignoreCastInfo) muxAddr._2 else muxAddr._2 + i
-      val muxBase = p.broadcastRMux.accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).length
-      val xBarRBase = p.xBarRMux.accessPars.length
-      val directRBase = p.directRMux.accessPars.length
-      val outputXBarRBase = p.xBarRMux.accessPars.sum
-      val outputDirectRBase = p.directRMux.accessPars.sum
-      val outputMuxBase = p.broadcastRMux.accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).sum
-      val vecId = if (ignoreCastInfo) i else castgrps.take(i).count(_ == castgrp)
-      if (bid == 0) {
-        Ledger.connectBroadcastR(this.hashCode, muxBase, vecId)
-        broadcastR(muxBase).connectLane(vecId,i,rBundle, backpressure)
-      }
-      Ledger.connectOutput(this.hashCode, outputXBarRBase + outputDirectRBase + outputMuxBase + vecId)
-      output(outputXBarRBase + outputDirectRBase + outputMuxBase + vecId)
-    }
-  }
-
-  def connectDirectWPort(wBundle: W_Direct, bufferPort: Int, muxAddr: (Int, Int))(implicit stack: List[KernelHash]): Unit = {
-    assert(p.hasDirectW)
-    assert(!usedMuxPorts.contains(("directW", (bufferPort,muxAddr._1,muxAddr._2,0,0))), s"Attempted to connect to directW port ($bufferPort,$muxAddr) twice!")
-    usedMuxPorts ::= ("directW", (bufferPort,muxAddr._1,muxAddr._2,0,0))
-    val bufferBase = p.directWMux.accessParsBelowBufferPort(bufferPort).length 
-    val muxBase = p.directWMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, muxAddr._2, 0).length
-    Ledger.connectDirectW(this.hashCode, bufferBase + muxBase)
-    directW(bufferBase + muxBase) := wBundle
-  }
-
-  def connectDirectRPort(rBundle: R_Direct, bufferPort: Int, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean)(implicit stack: List[KernelHash]): Seq[UInt] = {connectDirectRPort(rBundle, bufferPort, muxAddr, castgrps, broadcastids, ignoreCastInfo, true.B)}
-
-  def connectDirectRPort(rBundle: R_Direct, bufferPort: Int, muxAddr: (Int, Int), castgrps: List[Int], broadcastids: List[Int], ignoreCastInfo: Boolean, backpressure: Bool)(implicit stack: List[KernelHash]): Seq[UInt] = {
-    assert(p.hasDirectR)
-    castgrps.zip(broadcastids).zipWithIndex.map{case ((cg, bid), i) => 
-      val castgrp = if (ignoreCastInfo) 0 else cg
-      val effectiveOfs = if (ignoreCastInfo) muxAddr._2 else muxAddr._2 + i
-      val bufferBase = p.directRMux.accessParsBelowBufferPort(bufferPort).length
-      val xBarRBase = p.xBarRMux.accessPars.length
-      val muxBase = p.directRMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).length
-      val outputBufferBase = p.directRMux.accessParsBelowBufferPort(bufferPort).sum
-      val outputXBarRBase = p.xBarRMux.accessPars.sum
-      val outputMuxBase = p.directRMux(bufferPort).accessParsBelowMuxPort(muxAddr._1, effectiveOfs,castgrp).sum
-      val vecId = if (ignoreCastInfo) i else castgrps.take(i).count(_ == castgrp)
-      if (bid == 0) {
-        if (ignoreCastInfo && i == 0) {
-          assert(!usedMuxPorts.contains(("directR", (bufferPort,muxAddr._1,effectiveOfs,i,0))), s"Attempted to connect to directR port ($bufferPort,$muxAddr) twice!")
-          usedMuxPorts ::= ("directR", (bufferPort,muxAddr._1,effectiveOfs,i,0))
-        } else if (!ignoreCastInfo) {
-          assert(!usedMuxPorts.contains(("directR", (bufferPort,muxAddr._1,effectiveOfs,i,castgrp))), s"Attempted to connect to directR port ($bufferPort,$muxAddr) twice!")
-          usedMuxPorts ::= ("directR", (bufferPort,muxAddr._1,effectiveOfs,i,castgrp))
-        }
-        Ledger.connectDirectR(this.hashCode, bufferBase + muxBase, vecId)
-        directR(bufferBase + muxBase).connectLane(vecId,i,rBundle, backpressure)
-      }
-      Ledger.connectOutput(this.hashCode, outputXBarRBase + outputBufferBase + outputMuxBase + vecId)
-      output(outputXBarRBase + outputBufferBase + outputMuxBase + vecId)
-    }
-  }
-
-  def connectStageCtrl(done: Bool, en: Bool, port: Int): Unit = {
-    sEn(port) := en
-    sDone(port) := done
-  }
-
-
-}
 
 class NBufMem(p: NBufParams) extends Module {
